feat: Create LinkerPage and container and setup config at redux
