+++ linker-app/src/components/LinkerPage/index.ts
@@ -1,0 +1,2 @@
+import LinkerPage from './LinkerPage.container'
+export default LinkerPage
