+++ linker-app/src/modules/config/actions.ts
@@ -2,6 +2,7 @@ import { action, ActionType } from 'typesafe-actions'
 import { FETCH_CONFIG_REQUEST, FETCH_CONFIG_SUCCESS, FETCH_CONFIG_FAILURE, Config } from './types'
 
 export const fetchConfigRequest = () => action(FETCH_CONFIG_REQUEST, {})
+
 export const fetchConfigSuccess = (config: Config) => action(FETCH_CONFIG_SUCCESS, { config })
 export const fetchConfigFailure = (error: string) => action(FETCH_CONFIG_FAILURE, { error })
