+++ linker-app/src/sagas.ts
@@ -2,5 +2,5 @@ import { all } from 'redux-saga/effects'
 import { configSaga } from './modules/config/sagas'
 
 export function* rootSaga() {
-  yield all([configSaga])
+  yield all([configSaga()])
 }
